requirements.txt
python-binance>=1.0.19
python-telegram-bot>=20.7
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0
websockets>=11.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
loguru>=0.7.0
